pyproject.toml
[project]
name = "myspots"
version = "0.0.0"
description = "Search Google Maps for places and push to Airtable"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "click>=8.1.8",
    "fastkml>=1.1.0",
    "googlemaps>=4.10.0",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "networkx>=3.1",
    "pyyaml>=6.0.3",
    "textual>=1.0.0",
    "tqdm>=4.67.1",
    "ultimate-notion>=0.0.10",
]
authors = [
    { name = "Uri Laserson", email = "uri.laserson@gmail.com" }
]

[project.urls]
Homepage = "https://github.com/laserson/myspots"

[project.scripts]
myspots = "myspots.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
